+++ src/collective/themefragments/utils.py
@@ -1,13 +1,47 @@
 # -*- coding: utf-8 -*-
+from App.config import getConfiguration
 from ConfigParser import SafeConfigParser
-
 from plone.app.theming.interfaces import THEME_RESOURCE_NAME
+from plone.app.theming.interfaces import IThemingPolicy
 from plone.app.theming.plugins.utils import getPlugins
 from plone.resource.manifest import MANIFEST_FILENAME
+from zope.globalrequest import getRequest
 
 
-# This is copied from p.a.theming.plugins.utils to get uncached data
-# noinspection PyPep8Naming
+def cache(key):
+    def wrapper(func):
+        def cached(*args, **kwargs):
+            if getConfiguration().debug_mode:
+                return func(*args, **kwargs)
+
+            request = getRequest()
+            policy = IThemingPolicy(request)
+            cache_ = policy.getCache()
+            if not hasattr(cache_, 'collective.themefragments'):
+                setattr(cache_, 'collective.themefragments', {})
+            cache_ = getattr(cache_, 'collective.themefragments')
+
+            if callable(key):
+                key_ = key(*args)
+            else:
+                key_ = key
+
+            if key_ not in cache_:
+                cache_[key_] = func(*args, **kwargs)
+            return cache_[key_]
+
+        return cached
+    return wrapper
+
+
+@cache(lambda directory, section: ':'.join([directory.__name__, section]))
+def getFragmentsSettings(themeDirectory, section):
+    return getPluginSettings(
+        themeDirectory, plugins=[(section, None)]
+    ).get(section) or {}
+
+
+# This is copied from p.a.theming.plugins.utils to get properly cached data
 def getPluginSettings(themeDirectory, plugins=None):
     """Given an IResourceDirectory for a theme, return the settings for the
     given list of plugins (or all plugins, if not given) provided as a list